+++ src/main/java/plugily/projects/buildbattle/arena/impl/SoloArena.java
@@ -515,6 +515,11 @@ public class SoloArena extends BaseArena {
           User user = getPlugin().getUserManager().getUser(player);
           votingPlot.setPoints(votingPlot.getPoints() + user.getStat(StatsStorage.StatisticType.LOCAL_POINTS));
           user.setStat(StatsStorage.StatisticType.LOCAL_POINTS, 3);
+          if(!player.getInventory().contains(plugin.getVoteItems().getReportItem())) {
+            player.getInventory().addItem(plugin.getVoteItems().getReportItem());
+            player.getInventory().setItem(plugin.getVoteItems().getReportVoteItem().getSlot(), plugin.getVoteItems().getReportVoteItem().getItemStack());
+            player.updateInventory();
+          }
         }
       }
       if(!votingPlot.getOwners().isEmpty() && getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.ANNOUNCE_PLOTOWNER_LATER)) {